pyproject.toml
[tool.poetry]
name = "untexifyweb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.12"
# probably replace this with just 'tensorflow' on windows
tensorflow = "2.12.*"
numpy = "*"
matplotlib="*"
pillow="*"
opencv-python="*"
albumentations="*"
django = "^4.2.3"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
#FIXME: You don't need two installs of pyright.
pyright = "^1.1.318"
django-bootstrap5 = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
